Move auth.dart
